requirements.txt
colorama==0.4.6
flask==3.0.0
flask-cors==3.0.10
google-cloud-storage==2.9.0
numpy==1.22.0
PyYAML==6.0.1
requests==2.28.0
tqdm==4.66.1
halo==0.0.31
pandas==2.0.1
tensorflow==2.13.1
pillow==9.5.0
scikit-learn==1.3.1

